module8-task2 Добавлен вывод комментариев по нажатию кнопки
