+++ js/big-picture.js
@@ -1,27 +1,9 @@
 import {isEscapeKey} from './util.js';
+import {modifyCommentList} from './comments.js';
 
 const bodyElement = document.querySelector('body');
 const bigPictureElement = document.querySelector('.big-picture');
 const closeBigPictureElement = bigPictureElement.querySelector('#picture-cancel');
-const commentListElement = bigPictureElement.querySelector('.social__comments');
-const commentItemTemplate = commentListElement.querySelector('.social__comment');
-const commentListFragment = document.createDocumentFragment();
-
-/**
- * Функция выводит список комментариев пользователей
- * @param {Array} commentsData - Список комменатриев к фотографии
- */
-const modifyCommentList = (commentsData) => {
-  commentListElement.innerHTML = '';
-  commentsData.forEach((commentData) => {
-    const commentItemElement = commentItemTemplate.cloneNode(true);
-    commentItemElement.querySelector('.social__picture').setAttribute('src', commentData.avatar);
-    commentItemElement.querySelector('.social__picture').setAttribute('alt', commentData.name);
-    commentItemElement.querySelector('.social__text').textContent = commentData.message;
-    commentListFragment.append(commentItemElement);
-  });
-  commentListElement.append(commentListFragment);
-};
 
 /**
  * Функция выводит большое изображение на основе полученных данных
@@ -31,8 +13,6 @@ const modifyBigPicture = (photoData) => {
   bigPictureElement.querySelector('.big-picture__img img').setAttribute('src', photoData.url);
   bigPictureElement.querySelector('.likes-count').textContent = photoData.likes;
   bigPictureElement.querySelector('.social__caption').textContent = photoData.description;
-  bigPictureElement.querySelector('.social__comment-count').classList.add('hidden');
-  bigPictureElement.querySelector('.comments-loader').classList.add('hidden');
   modifyCommentList(photoData.comments);
 };
